--- routes/routes.js
@@ -1,35 +1,0 @@
-const router = require('express').Router();
-const {
-  getUsers, getUser, createUser, updateUser, updateAvatar,
-} = require('../controllers/users');
-const {
-  getCards, getCard, createCard, deleteCard, likeCard, dislikeCard,
-} = require('../controllers/cards');
-
-const mainPage = (req, res) => {
-  res.send({ message: 'Здесь будет главная страница' });
-};
-
-const errorPage = (req, res) => {
-  res.status(404);
-  res.send({ message: 'Запрашиваемый ресурс не найден' });
-};
-
-router.get('/', mainPage);
-router.get('/users', getUsers);
-router.get('/users/:userId', getUser);
-router.post('/users', createUser);
-router.patch('/users/me', updateUser);
-router.patch('/users/me/avatar', updateAvatar);
-
-
-router.get('/cards', getCards);
-router.get('/cards/:cardId', getCard);
-router.post('/cards', createCard);
-router.delete('/cards/:cardId', deleteCard);
-router.put('/cards/:cardId/likes', likeCard);
-router.delete('/cards/:cardId/likes', dislikeCard);
-
-router.use('*', errorPage);
-
-module.exports = router;
